raidboss: add basic triggers for Angra Mainyu (#3614)
* raidboss: add basic triggers for Angra Mainyu

Angra Mainyu is the first boss of The World of Darkness alliance raid.
This boss has a few instant-death mechanics, including some basic
mechanics that are very easy to add triggers for.

Add a trigger for the "Mortal Gaze" mechanic. When this is cast, players
should look away. If they fail to look away, they acquire the Doom
debuff and will die in 10 seconds unless they reach a cleansing circle.
Add an alert for the look away, and an alarm for when doom is acquired.

Add a trigger for the "Double Vision" attack. This attack is cast
periodically, and damages players based on their debuff. The whole arena
lights up half red and half white. If the player is hit by the red area
they gain Ireful, and if they are hit by the white area they gain
Sullen. Repeated hits of the same color do increasing damage until it is
fatal. Keep track of the current debuff and warn the player of which
color to stand in based on this.

In the future, it would be nice to add some alerts for Level 100 Flare,
Level 150 Doom, and possibly the roulette mechanic. These are more
complex and have been skipped for now.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* raidboss: remove trigger when nothing to do for double vision

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* raidboss: reword triggers for Double Vision

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* raidboss: use mainTank to decide Double Vision

If the player doesn't currently have one of the two debuffs, use the
status of whether they're the main tank or not for telling them if they
should stand in front or behind of Angra Mainyu.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>

* raidboss: remove unnecesasry mainTank from Angra Mainyu

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>