oopsy: remove SetInCombat from MistakeObserver api
The live list cares about in game combat for UI purposes
(hiding/showing more of the live list), but not for functional
purposes of tracking/resetting mistakes.

Have the live list itself subscribe to this event and remove
the in game combat event from MistakeObserver.

In the future, MistakeObserver will emit a sequence of events with
log timestamps (where possible) to make it easier to then synchronize
the summary page with a longer-running live list.

To get there, we need to clean up functions that happen only from
events.