oopsy: refactor combat timing and pulls out of collector
The main logic change here is that we are no longer using ACT
combat starting, and instead are tracking combat starting
based on actions / wipes / zone changes / in game combat state.
This will make it easier to add offline oopsy parsing in the
future.

This is also a refactoring to try to split up the responsibilities of
some of the classes. There was previously a dance between
DamageTracker, EffectTracker, and MistakeCollector about when pulls
started.

This patch adds CombatState which does the combat timing part
that MistakeCollector used to do.  It's a private member of
DamageTracker, and is tightly coupled to DamageTracker making
calls into it in order to decide when combat has started or not.

(If we split DamageTracker into LogProcessor and MistakeGenerator
then CombatState would be owned by LogProcessor, LogProcessor
would make calls into CombatState, and CombatState would make calls
into MistakeGenerator.)

Additionally, the early pull mistake parts of MistakeCollector are
now moved into DamageTracker.

This creates a more linear data flow of events ->
DamageTracker(+CombatState) -> EffectTracker -> MistakeCollector.

In the future, DamageTracker can be split into LogProcessor and
MistakeGenerator.  EffectTracker can be renamed PlayerStateTracker
(which generates internal special mistakes like missed buffs and
deaths).  MistakeCollector is more MistakeEventSyncerForwarder,
and in the future can then handle broadcast/sync to get the
summary page up to date.