oopsy: change MistakeObserver to be timestamp/event-based
This will allow for a separate observer to easily store an array
of ViewEvents.  Then a followup patch can broadcast a request
for these events and then receive them.

Timestamps in the views are now based on these events rather than
Date.now().  This will allow synced events to work properly and
will be (more?) consistent with death report timing.