raidboss: add alerts for Chimera in Cutter's Cry
Add basic alerts for the "Dragon's Voice" and "Ram's Voice" attacks of
the Chimera in Cutter's Cry.

These are straightforward alerts to warn the player to get in or get out
depending on if it is Dragon's Voice (A lightning donut AOE) or Ram's
Voice (An ice explosion in melee area).

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>