oopsy: synchronize events
MistakeCollector adds itself as an observer and collects all of the
ViewEvents.

When a new MistakeCollector is created, it does a RequestSync, which
broadcasts to any existing MistakeCollectors to send their events.
If one exists, they will SendSyncResponse with all the events along
with the time that it started collecting.

The MistakeCollector keeps track of its creation time and the time
of any SyncResponses that it has received so that it does not clobber
an earlier sync with a later sync.

One follow-up to this use of BroadcastEvent would be to have
PullCounter broadcast its number so that all pull counts could be
the same.  See the TODO about this in oopsy_summary_list.ts.