util: coinach add helper function that sorts keys
The original approach that calls
`JSON.stringify(data, Object.keys(data).sort(), 2)` works well
 in some situation, but it would
only sort the first dimention of the object,
and the worst thing is that it would drop nested object.
For example, call the function above with a object
like `{ foo: { bar: 1 } }`, the actual output would be
'{"foo":{}}' that lost nested object.

So the ideal way is to recursively sort keys before passing
to `JSON.stringify` function.