resources: build Regexes with parseHelper (#3689)
This is a copy and modification of the netregexes version.
As all the Regexes now use the same fields as the NetRegexes,
they can use the same CactbotBaseRegExp helpers.

This does remove the "this field is a \y{Float}" sorts of matching,
but NetRegexes hasn't seemed to have missed that.  Additionally, in
testing, many of these fields ended up being potentially blank in
some edge cases, so it's better to treat them as arbitrary strings.