oopsy: add conditional broadcast debug printing
I have not been able to reproduce this very well but I have seen
opening the summary sometimes (rarely) not display anything.
It seems most likely to be some sort of timing interaction with
OverlayPlugin broadcast events, and so hopefully this will help
track it down if it happens again.