eslint: Update Fixer Node Argument Range
Update fixer node argument range to use the first argument's first range
value and the last argument's last range value instead of using start
and end as they don't seem to be returned as part of the node argument
object anymore. Loc is available, but that doesn't give the text
position in characters and instead returns an object of { line, col }.

Fixes #3754.