raidboss: improve Zenos "protean" tell
The English text for the Art of the Sword ability zenos uses is just
"Protean". This may be obvious to folks used to savage raids but it
generally means an AOE mechanic where players have to spread out to
avoid overlapping line attacks.

Include the "(spread out)" to help players who don't recognize what
"protean" means.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>