raidboss: fix p2s dissociation callout (#3999)
In my previous testing it was outputting East when it was West, but based on other findings it was actually outting always East as it outputs West always without this change. 682D is the unknown spellid that always occurs after 682C. I believe it is tied to the head here and may yield better results.