raidboss: p3s fledgling flights and death's toll triggers (#4005)
I did not see a 006F marker and this is assuming death's toll markers are the same ids. I got the id's using the function provided and comparing to video + log I had for directions.

This solves Death's Toll rather than just stating the count. There is also the possibility of merging the two Death's Toll triggers by having the condition be in the output and using the log of a gainsEffect to set the boolean.