raidboss: fix timeline translation of zone syncs
This would cause zone seal messages to not sync in other languages,
causing timelines that only start on zone seal to not begin.

The issue is that the positive lookbehind group is always lazy (?!)
and so we need to make sure that the regex doesn't eat the colon
in the capturing group.

This partially addresses #4102, and may address #3949.