raidboss: predict last 2 EmEx swords to give more time (#4121)
During Emerald Ex phase 2, the boss will use an attack called Tertius
Terminus Est. This attack spawns several swords which perform cross or
plus aoes in sequence. Players need to predict which spots will explode
last and move near them and away from the first explosions in order to
safely avoid the AOEs.

The current implementation does not prompt players anything until all
6 swords have dropped. This gives players little time to figure out
what is safe and where to move after the first aoe.

It is seven seconds, but this can still leave players rushing and
mis-reading the queue.

We know that swords will always spawn in all 4 corners (along with two
in the middle). We can predict the last two sword spots by intersecting
the first 4 with the expected full set. This allows us to give the same
information a few seconds earlier, which may give players enough time to
orient themselves.

Additionally, provide an alert when the initial Tertius Terminus Est
cast is started. Doing this allows players to be primed to look for the
swords earlier, and not be caught off guard.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>