raidboss: improve Manusya/Transmute alerts
During The Tower of Zot's first boss, Minduruva, cactbot has some tells
for the Transmute Fire III and Transmute Bio III delayed aoe attacks.

These trigger immediately when the Transmute begins casting, and warn
the player to run behind or under the orb that will have the Fire or Bio
attacks.

The boss will perform 2 transmutation of the elements before then using
a third element directly. If the boss uses Bio or Fire, the players must
first stack appropriately on the boss before running to the first or
second orb.

Since there is no tell for this, the tell for the orb mechanic can lead
players to run to the spot early, making it difficult to reach the
correct safe spot.

To correct this, we want to tell the player to first handle the mechanic
for the boss before handling the orb mechanic.

This is done by tracking which mechanic is under which orb. Note that
Minduruva never uses the same element twice during an attack, and as far
as I have noticed doesn't use both Fire and Bio for transmute at the
same time.

Instead of immediately providing info for "Get Behind Orb" or "Get Under
Orb", use a Map to track if its a Fire or Bio mechanic awaiting.

Then, once the boss uses the Manusya ability, alert the player. In the
case of Thunder and Blizzard, just alert about the orb mechanic that is
happening. If it is Fire or Bio, perform an alert which indicates to Get
Behind or Get Under first, followed by the appropriate orb mechanic. In
the event that there is no orb mechanic (beginning of the fight, or the
two orbs were blizzard and thunder), then just alert the player about
the Fire or Bio mechanic using the standard outputs.

In all cases, there is enough time to handle the appropriate Manusya
mechanic on the boss first before running to the orb, so the delay in
the tell shouldn't cause an issue.

With this change we no longer warn the player to run away from the boss
too early, and alert them to stand near the boss to handle the on-boss
mechanic first.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>