oopsy: fix p1 hot cold spell bugs (#4185)
The hot/cold spell value carried over from the first set of
intemperate and could throw a mistake on the first hot/cold
of the second set of intemperate.  Stop being clever and
just track lose/gain.  Lose always happens prior to gaining
the new one, so this will not clobber the value.