ci: Remove Version Number from Release Folder (#4187)
Remove the version number from the release folder such that the release
structure will be cactbot-version.zip which contains cactbot/cacbot as
the nested folders. This will make it so in the future the folder under
ACT/Plugins for any new installs will be cactbot without a version number.

Tested with my fork by updating the GitHub Actions as well as the
VersionChecker code to point to my fork directly.