config: show timeline in edit timeline section (#4236)
The follow-up step to this is to add some ui that will allow for adding
custom timeline entries, but to have that make any sense, I think it's
useful for the timeline itself to be visible (and translated), so that
the times for custom timeline entries can have context.

A simpler alternative to this is to make a link to the triggers branch
timeline, but I thought this was nicer, and easy enough to translate
in place like we do for the tools.

This doesn't (and won't) let the user edit the timeline at all.
Perhaps there's some way to make this more obvious?  However,
I didn't want to disallow selection as copy and paste of times
seemed useful.  Not sure how else to indicate that this is a display
and not an editable text field.

Additionally, this doesn't respect the renaming or hiding options (only
the language), as config options don't know about raidboss options
because of the way options are loaded in user_config.ts.  I think this
is fine-ish, but it's likely to be confusing and I'm not sure how to
make it better.