refactor to use overlayHandler
The startsUsing would get 5.7s, the length of the cast, at most, which really only gives about 3 seconds notice but players will already know before the beginning of the cast where to go based on the jumps the knights do at end of 63D3 Strength of the Ward cast from King Thordan. Need to determine delaySeconds value here. Also not sure if this same spell is cast later in the encounter.