add sanctity ward direction
Could also have a call out for 'light party' positions (east/west) when the boss goes untargetable.
Then which direction movement will be once it has been determined. To also be determined on timing of that callout and what it should state. The idea is something like left/right or clockwise/counterclockwise to indicate to player which direction to first dodge the dashes. It can be all relative to center of arena (left/right), or to the top down view of the floor (clock/counterclock).
May also want a 'move' to signify when to begin the movement, but that's typically aligned with the dash.