raidboss: DSR Strength and Sanctity additional triggers (#4362)
* raidboss: DRS Spiral Thrust Safe Spots

* switch source from regex to array

* lint

* more lint

* number to output

* Add parenthesis

* reduce delay down

* refactor to use overlayHandler

The startsUsing would get 5.7s, the length of the cast, at most, which really only gives about 3 seconds notice but players will already know before the beginning of the cast where to go based on the jumps the knights do at end of 63D3 Strength of the Ward cast from King Thordan. Need to determine delaySeconds value here. Also not sure if this same spell is cast later in the encounter.

* fix lint warnings and errors

* remove console log message

* remove whitespace

* Add callout for boss location after dives

This also moves the matchedPositionToDir function globally as it would be used for multiple triggers

* remove capture false, fix typo in error message

* add sanctity ward direction

Could also have a call out for 'light party' positions (east/west) when the boss goes untargetable.
Then which direction movement will be once it has been determined. To also be determined on timing of that callout and what it should state. The idea is something like left/right or clockwise/counterclockwise to indicate to player which direction to first dodge the dashes. It can be all relative to center of arena (left/right), or to the top down view of the floor (clock/counterclock).
May also want a 'move' to signify when to begin the movement, but that's typically aligned with the dash.

* missed adding data

* rename sanctityWardDirection to sanctityWardDir

* reduce delaySeconds and Janlennoux name

the Ward directions do not work, there are 3 Ser Janellnoux combatants, the other one I have to look at more I think there may be a name typed wrong or something.

* reorder deletes

* change filter to building new array

* lint

still trying to get Janlenoux position data

* sort Janlenoux combatants, reduce delay on dives

* adjust delayseconds for clock callout

* fix case where dir is 0

* reduce thrust delay

* increase thrust safe spot delay

1 in like 40 pulls got bad data.

* remove some comments

Not sure if this commit as is counts for Sanctity of the Ward left/right direction, player may need to swap clockwise/counterclockwise callout dependent on their strat. I don't know of any other conflicts.

* switch to 1 move by default

* Update ui/raidboss/data/06-ew/ultimate/dragonsongs_reprise_ultimate.ts

Co-authored-by: Adrienne Walker <enne@quisquo.us>

* Update ui/raidboss/data/06-ew/ultimate/dragonsongs_reprise_ultimate.ts

Co-authored-by: Adrienne Walker <enne@quisquo.us>

* Update ui/raidboss/data/06-ew/ultimate/dragonsongs_reprise_ultimate.ts

Co-authored-by: Adrienne Walker <enne@quisquo.us>

* Update Sanctity Ward Direction comment

* Add Adelphel Charge Start Direction

* adjust time and lint

* fix comment spellid

* adjust delaySeconds

I'll see about getting heading written into this next

* adjust adelphel charge delay

* this seems to work at 4.7

* lint

* more commentry from pr comments

* remove TODO.

* add 4dir, change to compare x for janlenoux

* fix missing output parameter

* lint

Co-authored-by: Adrienne Walker <enne@quisquo.us>