test/i18n: enforce regex translations at test time (#4433)
This is working towards #1306.

Previously, the find missing translations script found these (although
it apparently missed some) and would not block merging in new errors.
This PR moves the regex translation to the test code, so `npm run test`
and github workflows will complain if they are missing.  As there are
now no more errors in this department after this PR, we can ratchet this
fixed so the followup can proceed.

The next steps here are to add code to popup-text to use replaceSync to
do the translation (if there is no locale regexes like netRegexDe).  For
backwards compat, the locale regexes will always be valid (and will
override this translation) in case any user triggers are using them in
weird ways.  However, once that it done we can deprecate locale regexes
and remove them from the documentation.

* i18n: added more translationsyncs (#4436)
* i18n: Small fixes (#4442)
Co-authored-by: Akurosia Kamo <kamothegamer@web.de>