resources/test: consolidate translation functions
To avoid writing this code a third time for translating regexes in
popup-text.ts, move translation functions into translations.ts.

This allows the runtime timeline/regex and test code to share the
same functionality, which makes it more likely that the test code
is testing the correct thing and the same translations are
happening at runtime as they are in tests.

This is working towards #1306.

The final steps are:
* auto-translation netRegex fields when netRegexDe (etc) is unspecified
* deprecate and remove netRegexDe (etc)