raidboss: fix edited-but-empty strings from being skipped
This was broken in #4449 because I didn't realize that the empty
string check was important behavior, whoops.

If a user edits something in the config ui and then clears it,
then it will leave an empty string in the config json and
output strings overrides.  This needs to be skipped (and could
be cleaned up elsewhere, but this is a small bugfix for now).