oopsy: add file viewer
This allows you to drag and drop a network log file onto
an html page and get a list of all mistakes.

This is still a bit of a work in progress, especially party buffs
which need to figure out who is in the party and is currently
depending on events that don't exist.

There's also a lot of TODOs here to improve the UI.  The summary
page could probably use a lot of work as well, in terms of being
able to select certain zones/encounters, and also to have a
summary of mistakes by type.

Fixes #3429.