oopsy: use log lines instead of events for party tracking (#4499)
In order for the log file-based oopsy_viewer to know who is in the
party, the PartyList log lines need to be used instead of the
PartyChanged events.  For simplicity, don't support both and
switch even the live version over to the log-based tracking
so there's only one code path.