raidboss: DSR reverse p6 hallowed wings left/right
Looking at fflogs replays and the log from #4564,
it appears that the "left wing" and "right wing" comments
are backwards.

We could consider saying "north" and "south" here to be
absolute, but I think most folks will be looking at Hraesvelgr
to see the wing and head, so left/right is pretty natural.