raidboss: make empty string sound mean no sound
There's currently no way to disable sound for a trigger.
This makes `sound: ''` to mean don't play any sound.

It's possible but unlikely this will fix #4588 due to the
`DSR Wrath Thunderstruck Targets` trigger having `sound: ''` and
`tts: 'null'`, but I was unable to repro this situation.

This also intersects with the #4523 changes.