i18n: fix a mistake in cn (#4641)
It seems the mistake was inherited from the Japanese version as the messages in other languages are correct. One can easily tell, even without understanding Japanese, that if '白玉へ' means 'go to the white orb', then '赤玉へ' definitely doesn't mean what it's supposed to mean. In fact '赤玉へ' means 'go to the red orb', but my japanese is too poor to fix it, sorry.