util: Fix fflogs.ts pagination handling (#4703)
It's messy, but it does seem to work. Previously, if event data were paginated, the subsequent `callFFLogs()` was being called with the wrong parameter, leading to corrupted + incomplete data being returned.

Separate from that, `Object.assign` was *overwriting* the `events` property on `mergedData`, not concatenating it. This wasn't visible previously due to the pagination bug, but it meant that after fixing that, if data *was* paginated, we would get only the last page of events.

Please do criticize, this is me tossing it in quickly so we can actually use it for the new content.