raidboss: Pandaemonium 7 normal package (#4713)
Nothing much to say on this one. Once I figured out which abilities to
ignore the loops weren't too hard to identify.

The proximity marker on Immortals' Obol kills people a lot because it's
logarithmic falloff damage, but because it always damages players it's
tough to make anything from it.