raidboss: Add Sephirot Unreal (#4716)
Largely a matter of find/replace. I finished up the timeline and
triggers prior to an attempt at running it, and everything seemed to
work okay. The "change to Earth Shakers" that was mentioned in one of
the live letters is simply that the targets are now *also* clearly
marked with stretchy tethers that turn purple when extended far enough.

Oopsy is untested, but the IDs all match, so I'm comfortable releasing
it as it stands.