raidboss: P6N Initial triggers/oopsy (#4730)
The timeline for this encounter has ended up being significantly more obstinate than I would like, so quisquous graciously agreed to take over on that.

I couldn't find *anything* in the network log relating to how Polyominous Dark works. (The radiating AoEs from the totems.) All of the totems have identical NPC models, and the cast ID is the same for all of them. There are no status effects on any of them. All of them have a facing of 0.0 from beginning to end of the cast, whether or not they hit a target.