raidboss: p7s add Light of Life bigAoe
The cast on this one is strange, fflogs showing it is canceled, castTime appears to be ~5s, and the ability goes off 26s after start of cast. This adds a ~5s notice for the aoe.

Also to note, strats may vary, but arguably it could be earlier call out such as get middle, big aoe because as far as I am aware, all strats require changing platform between purgations.