raidboss: p7s Roots of Attis Triggers (#4781)
These provide insight into which platform players should choose as the Roots of Attis spell will be breaking bridges between platforms.

The second and third occurrence will certainly lead to deaths if the wrong platforms are chosen prior to the cast finishing. I'm not sure if they should be upgraded to an alert.