raidboss: p7s Egg Tracking + Forbidden Fruits 1,4,6-10 (#4793)
The eggs spawn after ability 7811 completes, but it takes time for the
data to be available. I tested with 0.5s and 1s, both seemed to work. It
may be available shorter than 0.5s, but latency wise 0.5s is safer than
a lower number.

Currently calls out Left/Right for Forbidden Fruit 1, then the two
platforms for Forbidden Fruit 6, and Left/Right or South for Forbidden
Fruit 7.