resources: build NetRegexes from helper in most cases
This will make it possible to have the log guide have regexes
for things that don't have built-in cactbot regexes automatically
by using this helper.

This also changes the default "error name" for NetRegexes
to match the NetMatches type in most cases, and at least be
capitalized in others.  This is cosmetic and only shows up in errors.