raidboss: fix entering combat resetting `data`
Calling `getDataObject` returns a `RaidbossData` that does not have
`initData` called on it.  Prior to this patch, when entering combat,
`data` would be set to a new `data` object but without `initData`
called on it.

This isn't a larger issue because if `resetWhenOutOfCombat` is
true, then `this.Reset()` is called again in `SetInCombat`, fixing
the issues.  This only is an issue for the s-rank hunts in Thav
and Mare that have this flag set.

To avoid this problem in the future, always have `getDataObject`
return a properly initialized object.  To fix the short-term
issue, update `this.data.inCombat` directly without calling
`getDataObject` (oops).