replace data.perfectionShort/Long
This utilizes data.concept more. This also requires use of structuredClone for some of the list manipulation since it utilizes the InitialConcept type.

getMergePlayer altered slightly to account for short/long text. The call is used in two different ways. In some cases we want just the letter of the player, in other cases we need the short or long equivalent of the player in which case we still need to keep separate lists or pull from concept the a list of long and short players to match with.