remove comment about duster handling in timeline
It should all be sufficient now with it only showing the correct results. The only way to make it more correct would be to add some Suds based branches and/or left/right branches from Squeaky Clean, which given the short duration of the mechanics and that it's a "dungeon" boss is deemed more complex than it needs to be for this instance.