raidboss: ASS add Banishment triggers (#5015)
During the final boss, Banishment is the mechanic that happens between
Infern Brand 3 and Infern Ward 4. I have seen multiple strategies for
this mechanic, some simpler than others.

It looked to me like there was an invisible debuff, B9A, that has its
count field that correlates to the portal animation that the player has,
so this can be used to make triggers for where to position to solve the
mechanic. I compared raid_emulator with positioning I have in a log and
created the mapping.

I also added two additional triggers for baiting the wards. I was not
sure they are both necessary, I know the first one can be a bit more
trickier as it has extra time before the bait and at least two players
are likely to have been teleported in such a way that they need to move
to not trigger an Infern Wave going in the wrong direction. The second
wave has much longer stun lock before it goes off leaving less time to
react.

Will be able to test this in game later.