raidboss: DSR p7 Gigaflare direction to run (#4677)
Calls out clockwise or counterclockwise based on the atan2 of
determinant and dot product of first and second gigaflare
NetworkStartsCasting locations. Negative angle should mean one
direction, positive angle should mean the other.

Possible we could also replace this with relative to boss since I
imagine most groups are facing the boss?

~~To be tested.~~ It's probably okay now. I only ran with it once and
didn't have a problem with in/out tts over this tts, but it's easy
enough to test out with raidsim or fflogs data and typescript sim.