raidboss: ASS add Rush of Might Triggers (#5005)
Utilizes OverlayPlugin to get more up to date data. Log data was
inaccurate, this data was good when I tested.

I tested earlier and the second part was always correct. First part did
not call due to wrong rounded y1 values, but that has been fixed so I
think it's good. Wording will take some getting used to and there may be
reason to make this relative callouts instead of current card/intercad
based callouts since it can be difficult to understand at first.

Calls out `(NW/NE/SW/SE), (East/West) of line #(1-3)`, and the move
North/South for the followup cleave. Some handling for the move mechanic
if data from trigger 1 is bad to just state move without the North/South
component.

To break it down a bit more:
First set of cleaves is always Southwest or Southeast, whichever mirage
bursts more. Then depending on the difference, we either need to move
closer to the middle of the arena or can be right behind the first
cleave, which basically amounts to 2 bursts vs 3 bursts and 3 bursts vs
2 bursts requiring the player be on the atypical side of the diagonal
lines through the room in order to make it in time to the followup
cleave's safe spot.

The mechanic then repeats on either Northwest or Northeast.