raidboss: ASS Infern Brand 5 Triggers (#5016)
This adds a set of triggers to call out mechanic sequence in the last
boss' Infern Brand 5 mechanic. This is similar to Infern Brand 2, but it
adds the baited wards from Banishment to the equation.

This PR essentially:
- Calls where to start for the mechanic (Middle / Out)
- Calls when to cut (Using Magic Vulnerability and checking of previous
cut lines as safety measures)
- Calls additional move for the 4rth cut person after the baits
- Calls Baiting of the Ward # (# is somewhat superflous information)
- Calls bait puddles

Limitations:
The main limitation I see with it is I do not have a known method to
determine which Infern Ward is active until 0.5s before it goes off, so
this utilizes a record of the first flame color and relies on players to
have cut the correct colors. It only changes the color output on the 1st
cut flame. Alternatively, we could just remove the color portion of the
trigger until something better is out, but I figure if the first flame
cut is wrong, it's a wipe so this better than nothing.

Code code probably be cleaned up a bit more, it's repeating a lot of
output blocks.

I've only run this through raid emulator.

![image](https://user-images.githubusercontent.com/5702579/198823258-8401a702-9efb-4527-9234-12d45604f097.png)