raidboss: ASS/S add Hells' Nebula trigger (#5031)
Hells' Nebula is an ability done by trash mobs between the first and
second boss in ASS and ASS(S) that reduces all players' HP to 1. There
is no follow-up Doom debuff or other such mechanic that requires
immediately topping up players' health, other than the next source of
raidwide damage which is approximately 10s later. This trigger provides
a heads-up callout for the HP reduction.

There was a question if a trigger for this ability was useful when
merging the initial ASS triggers, so this pull request is made in
isolation so it can be rejected if undesired.