eureka: use fate/CE lines instead of events (#4959)
This will help with:
https://github.com/OverlayPlugin/OverlayPlugin/issues/90

This names the field for FateDirector properly and also fixes a missing
ceKey field (!!!) for CEDirector in netlog_defs.

Additionally, because netlog_defs uses `thingId` (vs `thingID`) style
naming, name `fateId` that way and update all of the zone files to match
that as well for consistency.

In local testing, this seems to work just fine in Zadnor. The only
behavior change is that now fates will appear even if the player missed
them being added in the first place.