raidboss: improvements to ASS/S silkie triggers (#5036)
I've attempted to add some triggers for handling Fresh Puff 1 (3 puffs
in triangle formation) and Fresh Puff 2 (4 puffs with tethers).

This is very much a wip - I've tested parts of this with user triggers,
but it needs some in-game testing (which I can't do right now because
servers are down). I'm planning to add a trigger for Fresh Puff 3 (2
puffs with effects, 6 "washed" away). I realistically won't be able to
get back to this until Thursday at the earliest though, so I'm throwing
this up now in case someone else has a chance to play around with it or
comment before then.

For Fresh Puff 2, the predominant strat I've seen is
[this](https://i.imgur.com/1Rvv4AL.png). There is an uptime strat for
boss = ice, puffs = cardinals, but it requires coordinated movement
between the two players with lightning puffs. One option is to provide a
shorter alert for that scenario without a resolution suggested, but it
would be inconsistent with the rest of the trigger. And the rest of the
resolutions are pretty standard (or the only possible resolution). So I
tried to compromise by creating output strings that users could easily
modify for this scenario, but.... yeah, definitely open to feedback.