util: handle undefined params in make_timeline.ts
It seems that entirely unused params are `undefined`,
and params without a value (e.g. `-f` on its own) are `null`.

I'm not sure when this changed, but the types are wrong in
this file now.

However, even with this change, nothing is emitted,
so there is more work to be done.

Partially addresses #5073.