util: refactor how missed events are handled
Previously, this code did not seem to be working at all and would
not emit missed events I'd expect.

However, its approach was also to check if every event was hit
but an event can be hit once and then missed on the next loop,
and that's important to point out.

This changes the approach to emit missed events on each sync,
rather than trying to figure out missed events after the fact.