util: add fflogs support to test_timeline (#5089)
This also fixes a bug in test_timeline's `testLineEvents` where it was
not setting the time before processing a log line. Normally, time
advances linearly with a timer and ticks periodically, and this causes
the set of active syncs to be updated. For test timelines, there's no
such timer, so the current time needs to be set prior to processing the
log line or the set of active syncs will be stale.

This is not an issue for `-f` style logs, as there's usually enough log
lines to update the time appropriately. However, for fflogs that only
include boss abilities, this causes many incorrect stale syncs.

Additionally, it seems unnecessary to sync after every long line, which
might create incorrect timings. If the sync occurs because a regex
matches, then the time will be updated as needed.